создает класс для инициализации films list
